Fix processed data filenames for input file pattern matching (#173)
Insert an additional tag in processed data filenames that indicates whether
a file is generated in the "compile" step or "encode" step. Change filename
matching pattern in dataset util to use only "encoded" files during training
or evaluation. This change fixes the incorrect mixing of the 2 kinds of data
in the input of training/evaluation, which causes errors.